ao_gps, ao_log: Make conversions to 16-bit time log values explicit
All log formats record only the low 16-bits of the time value, make
conversions to those explicit to silence -Wconversion warnings.

+++ src/kernel/ao_gps_report_metrum.c
@@ -41,7 +41,7 @@ ao_gps_report_metrum(void)
 		ao_mutex_put(&ao_gps_mutex);
 
 		if ((new & AO_GPS_NEW_DATA) && (gps_data.flags & AO_GPS_VALID)) {
-			gps_log.tick = ao_gps_tick;
+			gps_log.tick = (uint16_t) ao_gps_tick;
 			gps_log.type = AO_LOG_GPS_POS;
 			gps_log.u.gps.latitude = gps_data.latitude;
 			gps_log.u.gps.longitude = gps_data.longitude;
@@ -63,7 +63,7 @@ ao_gps_report_metrum(void)
 
 		if ((new & AO_GPS_NEW_TRACKING) && (n = gps_tracking_data.channels)) {
 			gps_log.type = AO_LOG_GPS_SAT;
-			gps_log.tick = ao_gps_tick;
+			gps_log.tick = (uint16_t) ao_gps_tick;
 			i = 0;
 			for (c = 0; c < n; c++) {
 				svid = gps_tracking_data.sats[c].svid;
