+++ src/kernel/ao_log_motor.c
@@ -46,7 +46,7 @@ ao_log(void)
 		ao_sleep(&ao_log_running);
 
 	ao_log_data.type = AO_LOG_FLIGHT;
-	ao_log_data.tick = ao_sample_tick;
+	ao_log_data.tick = (uint16_t) ao_sample_tick;
 	ao_log_data.u.flight.ground_accel = ao_ground_accel;
 	ao_log_data.u.flight.ground_accel_along = ao_ground_accel_along;
 	ao_log_data.u.flight.ground_accel_through = ao_ground_accel_through;
@@ -64,7 +64,7 @@ ao_log(void)
 		/* Write samples to EEPROM */
 		while (ao_log_data_pos != ao_data_head) {
 			AO_TICK_TYPE tick = ao_data_ring[ao_log_data_pos].tick;
-			ao_log_data.tick = tick;
+			ao_log_data.tick = (uint16_t) tick;
 			if ((AO_TICK_SIGNED) (tick - next_sensor) >= 0) {
 				ao_log_data.type = AO_LOG_SENSOR;
 				ao_log_data.u.sensor.pressure = ao_data_motor_pressure(&ao_data_ring[ao_log_data_pos]);
@@ -86,7 +86,7 @@ ao_log(void)
 		if (ao_flight_state != ao_log_state) {
 			ao_log_state = ao_flight_state;
 			ao_log_data.type = AO_LOG_STATE;
-			ao_log_data.tick = ao_time();
+			ao_log_data.tick = (uint16_t) ao_time();
 			ao_log_data.u.state.state = ao_log_state;
 			ao_log_data.u.state.reason = 0;
 			ao_log_write(&ao_log_data);
